allow-table-extended-to-handle-more-than-2048-chars (#326)
<!-- Please review our pull request review process in CONTRIBUTING.md before your proceed. -->

Resolves [#326](https://github.com/databricks/dbt-databricks/pull/326)

<!---
  Include the number of the issue addressed by this PR above if applicable.
  
  Example:
    resolves #1234

  Please review our pull request review process in CONTRIBUTING.md before your proceed.
-->

### Description

This PR allows the adapter to handle strings of more than 2048 characters when running show tables extended in Databricks. This is currently preventing users from being able to run this command when using dbt if they have a large number of tables in any given schema. 

The way this is done is that it will pass the `*` in instances where the string length of the pipe separated tables is more than 2048 chars.

I have successfully tested both unit tests and `tox -e integration-databricks-sql-endpoint`.

### Checklist

- [x] I have run this code in development and it appears to resolve the stated issue
- [x] This PR includes tests, or tests are not required/relevant for this PR
- [x] I have updated the `CHANGELOG.md` and added information about my change to the "dbt-databricks next" section.